requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
anyio==4.12.1
APScheduler==3.11.2
asyncpg==0.31.0
attrs==25.4.0
brotli==1.2.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
codecov==2.1.13
colorama==0.4.6
coverage==7.13.4
ddgs==9.10.0
fake-useragent==2.2.0
freezegun==1.5.5
frozenlist==1.8.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
lxml==6.0.2
multidict==6.7.1
packaging==26.0
pluggy==1.6.0
primp==0.15.0
propcache==0.4.1
Pygments==2.19.2
pytest==9.0.2  # Ignored CVE-2025-71176 in IDE, review needed
pytest-asyncio==1.3.0
pytest-cov==7.0.0
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-telegram-bot==22.6
requests==2.32.5
ruff==0.15.0
six==1.17.0
socksio==1.0.0
soupsieve==2.8.3
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
yarl==1.22.0